Various minor adjustments
Even though the subject suggests it's minor, the game-play is affected
too, which might need additional tweaking.

The primary change was to be explicit about the amount of updates
per second, and assure the game-state is based on the delta-time, not
the update frequency.

Here is a listing of the changes in detail:

* ship movement is now based on the delta-time.
* bullet rate was reduced to 20/s. It shouldn't be higher than the
  designated updates per second in the window's event loop for
  consistent results.
* Assure enemys are not colliding with the player when spawned.
* add `reset()` method to be more explicit.
* explicit mention of desired updates-per-second and frames-per-second
  in the game window's event loop.
* renamed bullet `update()` arg to `units` as it is closer to how
  the argument is used.